+++ changelog.md
@@ -7,6 +7,7 @@ Uiua is not yet stable.
 ## 0.7.0 - 2023-12-??
 ### Language
 - An entire Uiua codebase is now compiled before it is executed, rather than compiling and executing line-by-line
+- [`&i`](https://uiua.org/docs/&i) can now only be used as the first function in a binding
 - [`repeat` `⍥`](https://uiua.org/docs/repeat) can no longer use a negative number of repetitions
 ### Interpreter
 - Add the `uiua build` command, which emits a `.uasm` bytecode file